+++ tests/app/rdev/applications/environments/EnvironmentDetectorTest.php
@@ -5,16 +5,32 @@
  * Tests the environment detector class
  */
 namespace RDev\Applications\Environments;
+use RDev\Applications\Environments\Hosts\Host;
+use RDev\Applications\Environments\Hosts\HostRegistry;
 
 class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
 {
+    /** @var EnvironmentDetector The detector to use in tests */
+    private $detector = null;
+    /** @var HostRegistry|\PHPUnit_Framework_MockObject_MockObject The registry to use in tests */
+    private $registry = null;
+
+    /**
+     * Sets up the tests
+     */
+    public function setUp()
+    {
+        $this->detector = new EnvironmentDetector();
+        $this->registry = $this->getMock(HostRegistry::class);
+    }
+
     /**
      * Tests that the default environment is production
      */
     public function testThatDefaultEnvironmentIsProduction()
     {
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::PRODUCTION, $detector->detect([]));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn([]);
+        $this->assertEquals(Environment::PRODUCTION, $this->detector->detect($this->registry));
     }
 
     /**
@@ -22,14 +38,14 @@ class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
      */
     public function testThisServerBeingDevelopmentEnvironment()
     {
-        $config = [
-            "development" => gethostname(),
-            "testing" => "8.8.8.2",
-            "staging" => "8.8.8.8",
-            "production" => "8.8.8.4"
+        $hosts = [
+            "development" => [new Host(gethostname(), false)],
+            "testing" => [new Host("8.8.8.2", false)],
+            "staging" => [new Host("8.8.8.8", false)],
+            "production" => [new Host("8.8.8.4", false)]
         ];
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::DEVELOPMENT, $detector->detect($config));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn($hosts);
+        $this->assertEquals(Environment::DEVELOPMENT, $this->detector->detect($this->registry));
     }
 
     /**
@@ -37,14 +53,14 @@ class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
      */
     public function testThisServerBeingProductionEnvironment()
     {
-        $config = [
-            "development" => "8.8.8.8",
-            "testing" => "8.8.8.2",
-            "staging" => "8.8.8.4",
-            "production" => gethostname()
+        $hosts = [
+            "development" => [new Host("8.8.8", false)],
+            "testing" => [new Host("8.8.8.2", false)],
+            "staging" => [new Host("8.8.8.4", false)],
+            "production" => [new Host(gethostname(), false)]
         ];
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::PRODUCTION, $detector->detect($config));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn($hosts);
+        $this->assertEquals(Environment::PRODUCTION, $this->detector->detect($this->registry));
     }
 
     /**
@@ -52,14 +68,14 @@ class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
      */
     public function testThisServerBeingStagingEnvironment()
     {
-        $config = [
-            "development" => "8.8.8.8",
-            "testing" => "8.8.8.2",
-            "staging" => gethostname(),
-            "production" => "8.8.8.4"
+        $hosts = [
+            "development" => [new Host("8.8.8", false)],
+            "testing" => [new Host("8.8.8.2", false)],
+            "staging" => [new Host(gethostname(), false)],
+            "production" => [new Host("8.8.8.4", false)]
         ];
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::STAGING, $detector->detect($config));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn($hosts);
+        $this->assertEquals(Environment::STAGING, $this->detector->detect($this->registry));
     }
 
     /**
@@ -67,14 +83,14 @@ class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
      */
     public function testThisServerBeingTestingEnvironment()
     {
-        $config = [
-            "development" => "8.8.8.8",
-            "testing" => gethostname(),
-            "staging" => "8.8.8.2",
-            "production" => "8.8.8.4"
+        $hosts = [
+            "development" => [new Host("8.8.8", false)],
+            "testing" => [new Host(gethostname(), false)],
+            "staging" => [new Host("8.8.8.2", false)],
+            "production" => [new Host("8.8.8.4", false)]
         ];
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::TESTING, $detector->detect($config));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn($hosts);
+        $this->assertEquals(Environment::TESTING, $this->detector->detect($this->registry));
     }
 
     /**
@@ -84,28 +100,13 @@ class EnvironmentDetectorTest extends \PHPUnit_Framework_TestCase
     {
         // Truncate the last character of the host
         $truncatedHost = substr(gethostname(), 0, -1);
-        $config = [
-            "development" => "8.8.8.8",
-            "testing" => "8.8.8.2",
-            "staging" => [
-                ["type" => "regex", "value" => "/^" . preg_quote($truncatedHost, "/") . ".*$/"]
-            ],
-            "production" => "8.8.8.4"
+        $hosts = [
+            "development" => [new Host("8.8.8", false)],
+            "testing" => [new Host("8.8.8.2", false)],
+            "staging" => [new Host("/^" . preg_quote($truncatedHost, "/") . ".*$/", true)],
+            "production" => [new Host("8.8.8.4", false)]
         ];
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::STAGING, $detector->detect($config));
-    }
-
-    /**
-     * Tests a config with a callback
-     */
-    public function testWithCallback()
-    {
-        $callback = function ()
-        {
-            return "staging";
-        };
-        $detector = new EnvironmentDetector();
-        $this->assertEquals(Environment::STAGING, $detector->detect($callback));
+        $this->registry->expects($this->any())->method("getHosts")->willReturn($hosts);
+        $this->assertEquals(Environment::STAGING, $this->detector->detect($this->registry));
     }
 }